Determine how many more drones we actually need
The warehouse that recently acquired a new drone is attempting to fulfill all previously failed orders using the new drone. Subsequently, we check for any outstanding orders. If there are any, we make another attempt to distribute the remaining deliveries among the drones. If a new drone is required, we purchase one for the warehouse in need. This process continues until all deliveries are fulfilled.